+++ rocketpy/plots/rocket_plots.py
@@ -104,20 +104,26 @@ class _RocketPlots:
         """
 
         try:
-            x_power_drag_on = self.rocket.power_on_drag.x_array
-            y_power_drag_on = self.rocket.power_on_drag.y_array
+            x_power_drag_on = self.rocket.power_on_drag_by_mach.x_array
+            y_power_drag_on = self.rocket.power_on_drag_by_mach.y_array
         except AttributeError:
             x_power_drag_on = np.linspace(0, 2, 50)
             y_power_drag_on = np.array(
-                [self.rocket.power_on_drag.source(x) for x in x_power_drag_on]
+                [
+                    self.rocket.power_on_drag_by_mach.get_value_opt(x)
+                    for x in x_power_drag_on
+                ]
             )
         try:
-            x_power_drag_off = self.rocket.power_off_drag.x_array
-            y_power_drag_off = self.rocket.power_off_drag.y_array
+            x_power_drag_off = self.rocket.power_off_drag_by_mach.x_array
+            y_power_drag_off = self.rocket.power_off_drag_by_mach.y_array
         except AttributeError:
             x_power_drag_off = np.linspace(0, 2, 50)
             y_power_drag_off = np.array(
-                [self.rocket.power_off_drag.source(x) for x in x_power_drag_off]
+                [
+                    self.rocket.power_off_drag_by_mach.get_value_opt(x)
+                    for x in x_power_drag_off
+                ]
             )
 
         _, ax = plt.subplots()