+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Attention: The newest changes should be on top -->
 
 ### Changed
 
+- ENH: Improve multi-variable drag compatibility, regular-grid handling, and related tests/docs [#927](https://github.com/RocketPy-Team/RocketPy/pull/927/changes)
 - ENH: replace if elif else chains with match statement [#921](https://github.com/RocketPy-Team/RocketPy/pull/921/changes)
 - ENH: Refactor Flight class to improve time node handling and sensor/controllers [#843](https://github.com/RocketPy-Team/RocketPy/pull/843)